+++ tests/ethercat/test_ethercat_network.py
@@ -4,8 +4,9 @@ except ImportError:
     pass
 import pytest
 
+from ingenialink.ethercat.dictionary import EthercatDictionary
 from ingenialink.ethercat.network import EthercatNetwork
-from ingenialink.exceptions import ILFirmwareLoadError, ILError
+from ingenialink.exceptions import ILError, ILFirmwareLoadError
 
 
 @pytest.mark.docker
@@ -84,3 +85,17 @@ def test_scan_slaves_raises_exception_if_drive_is_already_connected(connect_to_s
     with pytest.raises(ILError):
         net.scan_slaves()
     assert servo.slave.state_check(pysoem.PREOP_STATE) == pysoem.PREOP_STATE
+
+
+@pytest.mark.ethercat
+def test_scan_slaves_info(read_config):
+    net = EthercatNetwork(read_config["ethercat"]["ifname"])
+    slaves_info = net.scan_slaves_info()
+    dictionary = EthercatDictionary(read_config["ethercat"]["dictionary"])
+
+    assert len(slaves_info) > 0
+    assert read_config["ethercat"]["slave"] in slaves_info
+    assert slaves_info[read_config["ethercat"]["slave"]].product_code == dictionary.product_code
+    assert (
+        slaves_info[read_config["ethercat"]["slave"]].revision_number == dictionary.revision_number
+    )